Update mail.php
This mail.php file is no longer needed for my website. The reason I removed it's function because, GitHub maintains only static pages and to run a php file I needed a dynamic environment and I'm not willing to pay for website. So, i came with a website named 99Imbound  which provided me with a embedded form of my reqirement for Free.